chore: CJK terminators + style gate + fast tests
* report._SENTENCE_TERMINATORS: register U+3002 / U+FF01 / U+FF1F so
  _consolidate_short_paragraphs no longer collapses Chinese exec
  summaries. CJK regression tests added. Offline replay + real bench
  bench2 lift overall 52.19 -> 52.55 (new
  22-run high for qid=1) and read 48.74 -> 51.82 (+3.08).

* scripts/check_style_rules.py: new gate enforcing no raw CJK in
  sources, test_ names <= 4 underscore segments, and -h/--help on
  every scripts/ entrypoint (with --exclude for future CJK fixture
  dirs). Wired into make check and make quick-check. scripts/README.md
  merges the two scattered -h convention blurbs; agent.md §Test
  Function Naming cross-references the gate.

* Test runtime: three timeout tests were paying 3.03s / 1.36s / 1.00s
  on retry backoff for behaviour irrelevant to the assertion. Disable
  retries in SkillExecutor timeout test, inject zero-backoff policy in
  WebSearch provider-errors test, widen ReviewRequest.timeout_seconds
  from int to float so review-queue timeout test can use 0.05s. No SDK
  test > 1s remains; SDK+coverage wall drops 13.82s -> 9.49s.

+++ houyi/assurance/verification/config.py
@@ -20,7 +20,7 @@ class EscalationPolicy(BaseModel):
     notify_channels: list[str] = Field(
         default_factory=list, description="Notification channels (slack, email)"
     )
-    timeout_seconds: int = Field(300, description="Timeout for human review")
+    timeout_seconds: float = Field(300, description="Timeout for human review")
     default_action: str = Field(
         "fail", description="Default action on timeout: 'fail', 'approve', 'retry'"
     )
